+++ src/back/zkif/int_ops.rs
@@ -7,15 +7,15 @@ use zkinterface_bellman::{
 use super::{
     zkif_cs::{fr_from_unsigned, ZkifCS},
     num::{Num, boolean_lc},
-    int32::Int32,
+    int64::Int64,
     bit_width::BitWidth,
 };
 
 pub fn bitwise_xor<E: Engine, CS: ConstraintSystem<E>>(
     cs: CS,
-    left: &Int32,
-    right: &Int32,
-) -> Int32
+    left: &Int64,
+    right: &Int64,
+) -> Int64
 {
     left.xor(cs, right).unwrap()
 }
@@ -23,9 +23,9 @@ pub fn bitwise_xor<E: Engine, CS: ConstraintSystem<E>>(
 // TODO: Implement directly on the type but fields are private.
 pub fn bitwise_and<E: Engine, CS: ConstraintSystem<E>>(
     mut cs: CS,
-    left: &Int32,
-    right: &Int32,
-) -> Int32
+    left: &Int64,
+    right: &Int64,
+) -> Int64
 {
     let out_bits: Vec<Boolean> =
         left.bits.iter()
@@ -34,14 +34,14 @@ pub fn bitwise_and<E: Engine, CS: ConstraintSystem<E>>(
                 Boolean::and(&mut cs, l, r).unwrap()
             ).collect();
 
-    Int32::from_bits(&out_bits)
+    Int64::from_bits(&out_bits)
 }
 
 pub fn bitwise_or<E: Engine, CS: ConstraintSystem<E>>(
     mut cs: &mut CS,
-    left: &Int32,
-    right: &Int32,
-) -> Int32
+    left: &Int64,
+    right: &Int64,
+) -> Int64
 {
     let out_bits: Vec<Boolean> =
         left.bits.iter()
@@ -50,7 +50,7 @@ pub fn bitwise_or<E: Engine, CS: ConstraintSystem<E>>(
                 bool_or(&mut cs, l, r)
             ).collect();
 
-    Int32::from_bits(&out_bits)
+    Int64::from_bits(&out_bits)
 }
 
 pub fn bool_or<'a, E, CS>(
@@ -68,10 +68,10 @@ pub fn bool_or<'a, E, CS>(
 pub fn div<E: Engine, CS: ConstraintSystem<E>>(
     mut cs: CS,
     numer_num: &Num<E>,
-    numer_int: &Int32,
+    numer_int: &Int64,
     denom_num: &Num<E>,
-    denom_int: &Int32,
-) -> (/*quotient*/ Num<E>, Int32, /*rest*/ Num<E>, Int32) {
+    denom_int: &Int64,
+) -> (/*quotient*/ Num<E>, Int64, /*rest*/ Num<E>, Int64) {
     let (quot_val, rest_val) = match (numer_int.value, denom_int.value) {
         (Some(numer), Some(denom)) => {
             if denom == 0 {
@@ -86,8 +86,8 @@ pub fn div<E: Engine, CS: ConstraintSystem<E>>(
         _ => (None, None)
     };
 
-    let quot_int = Int32::alloc(&mut cs, quot_val).unwrap();
-    let rest_int = Int32::alloc(&mut cs, rest_val).unwrap();
+    let quot_int = Int64::alloc(&mut cs, quot_val).unwrap();
+    let rest_int = Int64::alloc(&mut cs, rest_val).unwrap();
     // TODO: optimize the integer sizes.
 
     let quot_num = Num::from_int::<CS>(&quot_int);
@@ -103,7 +103,7 @@ pub fn div<E: Engine, CS: ConstraintSystem<E>>(
 
     // Verify that rest < denom.
     let diff_num = rest_num.clone() - denom_num;
-    let diff_int = Int32::from_num(&mut cs, &diff_num);
+    let diff_int = Int64::from_num(&mut cs, &diff_num);
     let ok = diff_int.is_negative();
     let one = CS::one();
     cs.enforce(
